Simplified Recommendation into single RecommendationController
